Added ability to check battery charging status to the smartplane service
Summary:
The smartplane/chargestatus field is set to 0 when battery is not charging, and changes to 1 when charging.
This patch adds a method updateChargingStatus() that can be called to check current battery charging status.
Also added is a delegate to the BLESmartplaneService class to subsequently receive charging status.

Test Plan: Cross fingers and hope it works. Untested.

Reviewers: samit.vaidya

Reviewed By: samit.vaidya

Differential Revision: https://office.tobyrich.com/D6